Account wiper (#153)
Add `DataProvider#getStripAndMergeAccountTransaction`, a
function that makes it easier to merge accounts with no balances but
trustlines, outstanding offers, etc. Returns a transaction to strip and merge
the account into another.

Right now, the function requires an account with zeroed non-native balances. 
In the future, it could auto-path-payment all the users funds.

Other changes:

- Upgraded `stellar-sdk` to 5.0.0.
- [DataProvider] Constructor now requires a `networkPassphrase` param.