doc update: move asset checker out from the approval provider (#197)
What

Besides fixing problems suggested by the grammarly pluggin, this PR moves the check for whether a regulated asset is involved out from the Approval Provider class.

Why

Checking whether an asset is a regulated asset doesn't have to be done by the approval provider. That is, one doesn't have to look at the information in the toml file in order to tell. It can be done by looking at the issuer's authorization flags.