doc update: remove asset from the constructor of approval provider (#198)
What

This PR adds two helper functions for getting the approval server once a regulated asset is found in a transaction. We first need to find the home domain of the asset, then we need to fetch the stellar.toml file and find the asset in the CURRENCIES section.

This PR also removes the regulatedAsset from the constructor of the approval provider.

Why

Once the corresponding approval server is found, we don't really have to provide the regulated asset in order to get an approval. One will need to send the whole transaction to the approval server.