doc update: support only single regulated asset in a transaction
What

This PR improves the helper function to return the regulated asset found rather than a boolean.

Why

The other helper function getHomeDomainByAssetIssuer needs to use the asset issuer's information in order to find the associated stellar.toml file. We can avoid having to iterate through the transaction object in a regulated asset is found.

At this moment we have decided to support only one regulated asset in the wallet sdk to avoid further complication.