plan update: simplify SEP-8 helper functions (#205)
What

This PR simplifies the helper functions planned by removing a redundant one.

Why

When we check whether an asset is regulated via Horizon, we already need to make the /account call to check the flags. Home domain information will be included in the account object. This change avoids one extra round trip to Horizon.